[pkg] fix: drop dependency on ca-certificates-java
Related to https://github.com/ev3dev-lang-java/ev3dev-lang-java/issues/728#issuecomment-633680662

The problem is that ca-certificates-java relies on jri-N-ev3
for the java-command; however jri-N-ev3 also depends on
ca-certificates java.

The proper fix would be to specify a self-dependency
as outlined in https://github.com/AdoptOpenJDK/openjdk-installer/issues/105#issuecomment-493939825 .
This will probably somehow break the dependency cycle
so that jri-N-ev3 is installed first and only then
the ca-certificates-java package gets installed.

This is nice, but the installation failure damages the system cacerts
file and simple --reinstall is *not* enough to fix this,
instead AFAIK all java packages have to be uninstalled and
then reinstalled again.

Let's instead ship the cacerts that ship with other AdoptOpenJDK builds.
This will fix cacerts error from the start. If the user wants to use
ca-certificates-java, this actually makes it possible to do so -
they just need to install the package after jri-N-ev3 and then
symlink /usr/lib/jvm/jri-N-ev3/lib/security/cacerts to
/etc/ssl/certs/java/cacerts which is managed by ca-certificates-java.

Nothing is perfect, though. The potential issue with Adopt's
CA certificates is that they seem pretty old (~ 3 years).
It seems that this is going to change: https://github.com/AdoptOpenJDK/openjdk-support/issues/13
This change will however impact the JDK build script and
not the packaging script.