[pkg] fix: add jri-N-ev3 self-dep to fix ca-certificates-java
With this, new clean installations should have working system-wide
java ca certificates. This is a non-invasive alternative to
https://github.com/ev3dev-lang-java/openjdk-ev3/pull/61 + https://github.com/ev3dev-lang-java/openjdk-ev3/pull/62.

The fix is likely based on a trick for making dpkg configure
this package first before configuring other packages that
depend on it.

The downside of this is that current installations will not get fixed.
They will still potentially have corrupted trust store from previous
failed installations. I don't know what the state of ev3dev is, but
if the package installation fails even in the ev3dev base image,
then *all* EV3 installations will have corrupted trust store.

The trust store can be un-corrupted by completely removing and
reinstalling ca-certificates-java. This is quite a heavy operation
for EV3, so this is a reason why the other solution might be preferred.